Move SlotMap to a separate file.
This commit doesn't contain any functional change. Just separate 2
SlotMap into a separate file, since it's a pretty large piece of logic.
